+++ advanced_alchemy/repository/_async.py
@@ -312,7 +312,7 @@ class SQLAlchemyAsyncRepositoryProtocol(FilterableRepositoryProtocol[ModelT], Pr
         *filters: Union[StatementFilter, ColumnElement[bool]],
         auto_expunge: Optional[bool] = None,
         statement: Optional[Select[tuple[ModelT]]] = None,
-        force_basic_query_mode: Optional[bool] = None,
+        count_with_window_function: Optional[bool] = None,
         error_messages: Optional[Union[ErrorMessages, EmptyType]] = Empty,
         load: Optional[LoadSpec] = None,
         execution_options: Optional[dict[str, Any]] = None,
@@ -422,6 +422,9 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
 
     This is useful for certain SQLAlchemy uses cases such as applying ``contains_eager`` to a query containing a one-to-many relationship
     """
+    count_with_window_function: bool = True
+    """Use an analytical window function to count results.  This allows the count to be performed in a single query.
+    """
 
     def __init__(
         self,
@@ -437,6 +440,7 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
         execution_options: Optional[dict[str, Any]] = None,
         wrap_exceptions: bool = True,
         uniquify: Optional[bool] = None,
+        count_with_window_function: Optional[bool] = None,
         **kwargs: Any,
     ) -> None:
         """Repository for SQLAlchemy models.
@@ -453,6 +457,7 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
             error_messages: A set of custom error messages to use for operations
             wrap_exceptions: Wrap SQLAlchemy exceptions in a ``RepositoryError``.  When set to ``False``, the original exception will be raised.
             uniquify: Optionally apply the ``unique()`` method to results before returning.
+            count_with_window_function: When false, list and count will use two queries instead of an analytical window function.
             **kwargs: Additional arguments.
 
         """
@@ -466,6 +471,9 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
         )
         self.wrap_exceptions = wrap_exceptions
         self.uniquify = self._get_uniquify(uniquify)
+        self.count_with_window_function = (
+            count_with_window_function if count_with_window_function is not None else self.count_with_window_function
+        )
         self._default_loader_options, self._loader_options_have_wildcards = get_abstract_loader_options(
             loader_options=load if load is not None else self.loader_options,
             inherit_lazy_relationships=self.inherit_lazy_relationships,
@@ -1469,7 +1477,7 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
         *filters: Union[StatementFilter, ColumnElement[bool]],
         statement: Optional[Select[tuple[ModelT]]] = None,
         auto_expunge: Optional[bool] = None,
-        force_basic_query_mode: Optional[bool] = None,
+        count_with_window_function: Optional[bool] = None,
         order_by: Optional[Union[list[OrderingPair], OrderingPair]] = None,
         error_messages: Optional[Union[ErrorMessages, EmptyType]] = Empty,
         load: Optional[LoadSpec] = None,
@@ -1483,7 +1491,7 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
             *filters: Types for specific filtering operations.
             statement: To facilitate customization of the underlying select query.
             auto_expunge: Remove object from session before returning.
-            force_basic_query_mode: Force list and count to use two queries instead of an analytical window function.
+            count_with_window_function: When false, list and count will use two queries instead of an analytical window function.
             order_by: Set default order options for queries.
             error_messages: An optional dictionary of templates to use
                 for friendlier error messages to clients
@@ -1495,12 +1503,15 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
         Returns:
             Count of records returned by query, ignoring pagination.
         """
+        self.count_with_window_function = (
+            count_with_window_function if count_with_window_function is not None else self.count_with_window_function
+        )
         self.uniquify = self._get_uniquify(uniquify)
         error_messages = self._get_error_messages(
             error_messages=error_messages,
             default_messages=self.error_messages,
         )
-        if self._dialect.name in {"spanner", "spanner+spanner"} or force_basic_query_mode:
+        if self._dialect.name in {"spanner", "spanner+spanner"} or count_with_window_function:
             return await self._list_and_count_basic(
                 *filters,
                 auto_expunge=auto_expunge,
@@ -1678,7 +1689,12 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
         execution_options: Optional[dict[str, Any]],
     ) -> Select[tuple[int]]:
         # Count statement transformations are static
-        return statement.with_only_columns(sql_func.count(text("1")), maintain_column_froms=True).order_by(None)
+        return (
+            statement.with_only_columns(sql_func.count(text("1")), maintain_column_froms=True)
+            .limit(None)
+            .offset(None)
+            .order_by(None)
+        )
 
     async def upsert(
         self,
@@ -1695,7 +1711,7 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
         execution_options: Optional[dict[str, Any]] = None,
         uniquify: Optional[bool] = None,
     ) -> ModelT:
-        """Update or create instance.
+        """Modify or create instance.
 
         Updates instance with the attribute values present on `data`, or creates a new instance if
         one doesn't exist.
@@ -1777,7 +1793,7 @@ class SQLAlchemyAsyncRepository(SQLAlchemyAsyncRepositoryProtocol[ModelT], Filte
         execution_options: Optional[dict[str, Any]] = None,
         uniquify: Optional[bool] = None,
     ) -> list[ModelT]:
-        """Update or create instance.
+        """Modify or create multiple instances.
 
         Update instances with the attribute values present on `data`, or create a new instance if
         one doesn't exist.
@@ -2161,20 +2177,20 @@ class SQLAlchemyAsyncQueryRepository:
     async def list_and_count(
         self,
         statement: Select[Any],
-        force_basic_query_mode: Optional[bool] = None,
+        count_with_window_function: Optional[bool] = None,
         **kwargs: Any,
     ) -> tuple[list[Row[Any]], int]:
         """List records with total count.
 
         Args:
             statement: To facilitate customization of the underlying select query.
-            force_basic_query_mode: Force list and count to use two queries instead of an analytical window function.
+            count_with_window_function: Force list and count to use two queries instead of an analytical window function.
             **kwargs: Instance attribute value filters.
 
         Returns:
             Count of records returned by query, ignoring pagination.
         """
-        if self._dialect.name in {"spanner", "spanner+spanner"} or force_basic_query_mode:
+        if self._dialect.name in {"spanner", "spanner+spanner"} or count_with_window_function:
             return await self._list_and_count_basic(statement=statement, **kwargs)
         return await self._list_and_count_window(statement=statement, **kwargs)